Refactor configuration lookups
This commit changes how configuration is looked up. Previously,
the database (db) was used as the lookup key. This worked, but it
wouldn't work if the user wanted to configure two servers that
have different databases named the same (for example: wordpress
on two different servers).

Now an arbitrary "connection" key is used.